+++ src/HalfEdge2D/Renderables/Point.cpp
@@ -3,7 +3,7 @@
 Point::Point()
 {
     m_Size = 0.01f;
-    m_Position = Vec2f(0.0f, 0.0f);
+    m_Position = Vec2d(0.0, 0.0);
     m_Colour= Vec4f(0.0f, 0.0f, 0.0f, 1.0f);
     m_Visible = true;
 }
@@ -18,7 +18,7 @@ const float& Point::getSize() const
     return m_Size;
 }
 
-const Vec2f& Point::getPosition() const
+const Vec2d& Point::getPosition() const
 {
     return m_Position;
 }
@@ -38,7 +38,7 @@ void Point::setSize(const float& size)
     m_Size = size;
 }
 
-void Point::setPosition(const Vec2f& pos)
+void Point::setPosition(const Vec2d& pos)
 {
     m_Position = pos;
 }
